Create booklist.txt
This file will create a stack of books by using a linked list. It will ask the user for the title and author of the book. The user will be allowed to print the list of books (top to bottom), remove a book from the top of the list, and allow the user to view the top book in the stack.